+++ ZetaRenderPass/Display/Display.hlsl
@@ -145,7 +145,7 @@ float4 mainPS(VSOut psin) : SV_Target
 	else if (g_local.DisplayOption == (int) DisplayOption::NORMAL)
 	{
 		GBUFFER_NORMAL g_normal = ResourceDescriptorHeap[g_frame.CurrGBufferDescHeapOffset + GBUFFER_OFFSET::NORMAL];
-		half2 encodedNormal = g_normal.SampleLevel(g_samPointClamp, uv, 0);
+		float2 encodedNormal = g_normal.SampleLevel(g_samPointClamp, uv, 0);
 		display = Math::Encoding::DecodeUnitNormal(encodedNormal.xy);
 		display = display * 0.5 + 0.5;
 	}
